Brute-forces `AutoComplete` transpilation to avoid race condition from Webpack's Copy plugin
