+++ packages/fuselage/webpack.config.js
@@ -2,7 +2,6 @@
 
 const path = require('path');
 
-const CopyPlugin = require('copy-webpack-plugin');
 const webpack = require('webpack');
 const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 
@@ -102,15 +101,6 @@ module.exports = (env, { mode = 'production' }) => ({
     '@rocket.chat/fuselage-hooks',
   ],
   plugins: [
-    new CopyPlugin({
-      patterns: [
-        {
-          context: 'src/',
-          from: '**/!(scss).d.ts',
-          to: '../dist',
-        },
-      ],
-    }),
     new webpack.DefinePlugin({
       'process.env.VERSION': JSON.stringify(pkg.version),
     }),
